[ADD] Image processing [REFACT] File organisation for an easier integration
