--- raspberryPi/CAN/inc/CAN_Application.h
@@ -1,7 +1,0 @@
-#ifndef CAN_APPLICATION_H
-#define CAN_APPLICATION_H
-
-void launchCANServices (void);
-int getControlCan(void);
-
-#endif
