--- raspberryPi/CAN/src_test/test_entire_application_withInterface.c
@@ -1,6 +1,0 @@
-#include "CAN_Application.h"
-
-int main(void){
-	launchCANServices();
-	return 0;
-}
