--- raspberryPi/Integration/CAN_Application.h
@@ -1,6 +1,0 @@
-#ifndef CAN_APPLICATION_H
-#define CAN_APPLICATION_H
-
-void launchCANServices (void);
-
-#endif
