+++ raspberryPi/include/Interface/PosVoitInterface.h
@@ -1,0 +1,9 @@
+#ifndef INTERFACES_VOIT_H
+#define INTERFACES_VOIT_H
+
+#include "traitement_image/traitementimg.hpp"
+
+void initCarPosition();
+
+positionVoiture* linkPositionVoiture();
+#endif
