+++ raspberryPi/src/StateMachine/Transition.cpp
@@ -1,0 +1,20 @@
+#include "StateMachine/Transition.h"
+
+Transition::Transition(int src, int dest, int (*f)())
+{
+  srcId=src;
+  destId=dest;
+  valuationFunction=f;
+}
+
+Transition::~Transition(){}
+
+int Transition::getSrc()
+{
+    return srcId;
+}
+
+int Transition::getDest()
+{
+    return destId;
+}
